Squashed commit for the login-and-register branch rebase to master
Login and Register setup with a single view, minor validation

Favicon magic

In loginview, register functionalities (some login) and verification

Cookies, mostly serverside stuff

"That's a whole lot of code": login, register and validation
Routeguard is working with Vuex to check for session validation, and
redirects to either login if not valid, or that what you're trying to
access if it is valid.

The login-and-register branch makes in one view the functionalities
work for login and register, but also checks if the cookie (session) is valid
and does some route magic in combination with vuex.